Remove dependency on rust
