--- .github/workflows/release.yml
@@ -1,83 +1,0 @@
-name: Build precompiled NIFs
-
-on:
-  push:
-    branches:
-      - main
-    tags:
-      - "*"
-
-jobs:
-  build_release:
-    name: NIF ${{ matrix.nif }} - ${{ matrix.job.target }} (${{ matrix.job.os }})
-    runs-on: ${{ matrix.job.os }}
-    strategy:
-      fail-fast: false
-      matrix:
-        nif: ["2.16", "2.17"]
-        job:
-          - {
-              target: aarch64-unknown-linux-gnu,
-              os: ubuntu-20.04,
-              use-cross: true,
-            }
-          - {
-              target: aarch64-unknown-linux-musl,
-              os: ubuntu-20.04,
-              use-cross: true,
-            }
-          - { target: aarch64-apple-darwin, os: macos-11 }
-          - {
-              target: riscv64gc-unknown-linux-gnu,
-              os: ubuntu-20.04,
-              use-cross: true,
-            }
-          - { target: x86_64-apple-darwin, os: macos-11 }
-          - { target: x86_64-unknown-linux-gnu, os: ubuntu-20.04 }
-          - {
-              target: x86_64-unknown-linux-musl,
-              os: ubuntu-20.04,
-              use-cross: true,
-            }
-          - { target: x86_64-pc-windows-gnu, os: windows-2019 }
-          - { target: x86_64-pc-windows-msvc, os: windows-2019 }
-
-    steps:
-      - name: Checkout source code
-        uses: actions/checkout@v3
-
-      - name: Extract project version
-        shell: bash
-        run: |
-          # Get the project version from mix.exs
-          echo "PROJECT_VERSION=$(sed -n 's/^  @version "\(.*\)"/\1/p' mix.exs | head -n1)" >> $GITHUB_ENV
-
-      - name: Install Rust toolchain
-        uses: dtolnay/rust-toolchain@stable
-        with:
-          toolchain: stable
-          target: ${{ matrix.job.target }}
-
-      - name: Build the project
-        id: build-crate
-        uses: philss/rustler-precompiled-action@v1.0.1
-        with:
-          project-name: uuidv7
-          project-version: ${{ env.PROJECT_VERSION }}
-          target: ${{ matrix.job.target }}
-          nif-version: ${{ matrix.nif }}
-          use-cross: ${{ matrix.job.use-cross }}
-          project-dir: "native/uuidv7"
-
-      - name: Artifact upload
-        uses: actions/upload-artifact@v3
-        with:
-          name: ${{ steps.build-crate.outputs.file-name }}
-          path: ${{ steps.build-crate.outputs.file-path }}
-
-      - name: Publish archives and packages
-        uses: softprops/action-gh-release@v1
-        with:
-          files: |
-            ${{ steps.build-crate.outputs.file-path }}
-        if: startsWith(github.ref, 'refs/tags/')
