--- checksum-Elixir.UUIDv7.exs
@@ -1,20 +1,0 @@
-%{
-  "libuuidv7-v0.2.1-nif-2.16-aarch64-apple-darwin.so.tar.gz" => "sha256:a49491f9d2d7c2e18de320db86de2b92d96df5f9f73765cab453a40b9ed6ed94",
-  "libuuidv7-v0.2.1-nif-2.16-aarch64-unknown-linux-gnu.so.tar.gz" => "sha256:52448e03a19dc56a1ae41c407e6a39e9fd3a2a860f0ee88f7ecab11cc53aca4d",
-  "libuuidv7-v0.2.1-nif-2.16-aarch64-unknown-linux-musl.so.tar.gz" => "sha256:3d58f44bc1f1be8bd1a6209164554ad6661494996a59e0f3182b534023906c95",
-  "libuuidv7-v0.2.1-nif-2.16-riscv64gc-unknown-linux-gnu.so.tar.gz" => "sha256:7c11b113a9a1bd62262889ea8c797a0e5660cab52a578a99fc0df6983ebbc95e",
-  "libuuidv7-v0.2.1-nif-2.16-x86_64-apple-darwin.so.tar.gz" => "sha256:4071cd2d0318949c56583cce7bb6f62b6627072ae0a4bc3917f1a8c2ed3c51ae",
-  "libuuidv7-v0.2.1-nif-2.16-x86_64-unknown-linux-gnu.so.tar.gz" => "sha256:699d005aa314bc0ae8e8e7d52731761985329999f5d1019136972e0b30661a8d",
-  "libuuidv7-v0.2.1-nif-2.16-x86_64-unknown-linux-musl.so.tar.gz" => "sha256:2cbcf6902fee866d79568f88977c9a9145476bf212be8c23af062ed67d3819ab",
-  "libuuidv7-v0.2.1-nif-2.17-aarch64-apple-darwin.so.tar.gz" => "sha256:8115e80cedded1722ed96ccb2f4898bc40c42b377fef26025ebb6370f824ee8a",
-  "libuuidv7-v0.2.1-nif-2.17-aarch64-unknown-linux-gnu.so.tar.gz" => "sha256:a256a6e61d13da15540797a5387259ea49526b9c71164be023fea4c82bcfa9ef",
-  "libuuidv7-v0.2.1-nif-2.17-aarch64-unknown-linux-musl.so.tar.gz" => "sha256:846fd1a22478d09a07a701a184f56e19b86fe55167f497e0ac84a997b8fdf3e3",
-  "libuuidv7-v0.2.1-nif-2.17-riscv64gc-unknown-linux-gnu.so.tar.gz" => "sha256:249e368f2f4f6cc6d08a20ed5435540117a7ace83471b009af792e7a01159e21",
-  "libuuidv7-v0.2.1-nif-2.17-x86_64-apple-darwin.so.tar.gz" => "sha256:ee216606432003abcc1ffe0611068d796649680693df8d65cd91b9ccd073afaa",
-  "libuuidv7-v0.2.1-nif-2.17-x86_64-unknown-linux-gnu.so.tar.gz" => "sha256:e4873ec5dfcc5510dee393c983a568f2a38f7cf3d2ee7ed7d45694de5eee4081",
-  "libuuidv7-v0.2.1-nif-2.17-x86_64-unknown-linux-musl.so.tar.gz" => "sha256:1b5b20fec37ce3c9f8ce6bbbe5476f5315dc90c3e7670c9c1e94578b569aa1bb",
-  "uuidv7-v0.2.1-nif-2.16-x86_64-pc-windows-gnu.dll.tar.gz" => "sha256:bdd448e7c387f874b12b79037e9e66056a114895a66c7e3b39a19f4d1d559b70",
-  "uuidv7-v0.2.1-nif-2.16-x86_64-pc-windows-msvc.dll.tar.gz" => "sha256:3f22edb7049fc056c861a6ea9f4fb0bc557b0fec594096c481b2fdb124ce9557",
-  "uuidv7-v0.2.1-nif-2.17-x86_64-pc-windows-gnu.dll.tar.gz" => "sha256:ceefb208d65fa86fac769ca58789b2b495433fd6005b3d3558f26b3ef42a4e49",
-  "uuidv7-v0.2.1-nif-2.17-x86_64-pc-windows-msvc.dll.tar.gz" => "sha256:c40b94bb04f96ab336940bdee722a18fb0e4ac0499209b97179b88867715f40a",
-}
