--- native/uuidv7/.cargo/config.toml
@@ -1,26 +1,0 @@
-[target.x86_64-apple-darwin]
-rustflags = [
-  "-C", "link-arg=-undefined",
-  "-C", "link-arg=dynamic_lookup",
-]
-
-[target.aarch64-apple-darwin]
-rustflags = [
-  "-C", "link-arg=-undefined",
-  "-C", "link-arg=dynamic_lookup",
-]
-
-# See https://github.com/rust-lang/rust/issues/59302
-[target.x86_64-unknown-linux-musl]
-rustflags = [
-  "-C", "target-feature=-crt-static"
-]
-
-# See https://github.com/rust-lang/rust/issues/59302
-[target.aarch64-unknown-linux-musl]
-rustflags = [
-  "-C", "target-feature=-crt-static"
-]
-
-[profile.release]
-lto = true
