--- native/uuidv7/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "uuidv7"
-version = "0.1.1"
-authors = []
-edition = "2021"
-
-[lib]
-name = "uuidv7"
-path = "src/lib.rs"
-crate-type = ["cdylib"]
-
-[dependencies]
-rustler = { version = "0.30", default-features = false, features = [
-  "derive",
-  "nif_version_2_16",
-] }
-uuid = { version = "1.6.1", features = ["v7"] }
