--- native/uuidv7/Cross.toml
@@ -1,4 +1,0 @@
-[build.env]
-passthrough = [
-  "RUSTLER_NIF_VERSION"
-]
